Add support for parameters
This commit adds support for recipe parameters. Those parameters can be passed when calling the `recipe` extension method, and will be passed to
the recipe instance, which can react to user supplied values. For example, the `git-commit` recipe now accepts an optional `baseUrl` which
allows linking to GitHub, for example.