Migrate configuration from `buildScan` to `buildScanRecipes`
This commit changes the configuration block of this plugin, to use its own
extension instead of using the build scan extension. This was done to prevent
some cryptic classloading issues to occur when multiple versions of the plugin
are applied, from an init script and the build script itself, typically.

So, instead of writing:

```
buildScan {
   recipe 'git-commit'
}
```

you'll need to write:

```
buildScanRecipes {
   recipe 'git-commit'
}
```